Conflicts:
	backend/src/Lighthouse/CoreBundle/Tests/Controller/OrderControllerTest.php

+++ webfront/web/blocks/form/form_order/form_order.js
@@ -26,12 +26,16 @@ define(function(require, exports, module) {
                     });
 
                 block.renderEditForm(orderProductModel);
+            },
+            'click .form_orderProduct__cancelLink': function(e) {
+                var block = this;
+                block.$tr.detach();
             }
         },
         initialize: function() {
             var block = this;
 
-            if (block.model.id){
+            if (block.model.id) {
                 document.getElementById('form_order__removeLink').addEventListener('click', function(e) {
                     e.preventDefault();
 
@@ -39,15 +43,16 @@ define(function(require, exports, module) {
                         return;
                     }
 
-                    e.target.classList.add('preloader_rows');
+                    if (confirm('Вы уверены?')) {
+                        e.target.classList.add('preloader_rows');
+                        block.disable();
 
-                    block.disable();
-
-                    block.model.destroy({
-                        success: function() {
-                            router.navigate('/orders');
-                        }
-                    });
+                        block.model.destroy({
+                            success: function() {
+                                router.navigate('/orders');
+                            }
+                        });
+                    }
                 });
             }
 
@@ -72,6 +77,10 @@ define(function(require, exports, module) {
                 el: Form_orderProduct.prototype.template({
                     model: orderProductModel
                 })
+            });
+
+            form.once('submit:success', function() {
+                block.$tr.detach();
             });
 
             block.$tr.find('td').html(form.el);
